third_party/blink/web_tests/platform/mac-mac12/external/wpt/css/css-overflow/scrollbar-gutter-vertical-lr-001-expected.txt

This is a testharness.js-based test.
[FAIL] overflow scroll, scrollbar-gutter auto
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow auto, scrollbar-gutter stable
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow scroll, scrollbar-gutter stable
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow hidden, scrollbar-gutter stable
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow auto, scrollbar-gutter stable both-edges
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow scroll, scrollbar-gutter stable both-edges
  assert_greater_than: content height expected a number greater than 200 but got 200
[FAIL] overflow hidden, scrollbar-gutter stable both-edges
  assert_greater_than: content height expected a number greater than 200 but got 200
Harness: the test ran to completion.
